refactor: use DDD folder structure
Add `import/no-default-export` to the project
